Moved everything related to text-editor to a dir
