+++ src/text-editor.js
@@ -22,7 +22,7 @@ const {
   isHalfWidthCharacter,
   isKoreanCharacter,
   isWrapBoundary
-} = require('./text-utils');
+} = require('./text-editor/utils');
 
 const SERIALIZATION_VERSION = 1;
 const NON_WHITESPACE_REGEXP = /\S/;
@@ -77,21 +77,21 @@ module.exports = class TextEditor {
 
   static setScheduler(scheduler) {
     if (TextEditorComponent == null) {
-      TextEditorComponent = require('./text-editor-component');
+      TextEditorComponent = require('./text-editor/component');
     }
     return TextEditorComponent.setScheduler(scheduler);
   }
 
   static didUpdateStyles() {
     if (TextEditorComponent == null) {
-      TextEditorComponent = require('./text-editor-component');
+      TextEditorComponent = require('./text-editor/component');
     }
     return TextEditorComponent.didUpdateStyles();
   }
 
   static didUpdateScrollbarStyles() {
     if (TextEditorComponent == null) {
-      TextEditorComponent = require('./text-editor-component');
+      TextEditorComponent = require('./text-editor/component');
     }
     return TextEditorComponent.didUpdateScrollbarStyles();
   }
@@ -5294,9 +5294,9 @@ module.exports = class TextEditor {
   getElement() {
     if (!this.component) {
       if (!TextEditorComponent)
-        TextEditorComponent = require('./text-editor-component');
+        TextEditorComponent = require('./text-editor/component');
       if (!TextEditorElement)
-        TextEditorElement = require('./text-editor-element');
+        TextEditorElement = require('./text-editor/element');
       this.component = new TextEditorComponent({
         model: this,
         updatedSynchronously: TextEditorElement.prototype.updatedSynchronously,
