+++ src/text-editor/utils.js
